move OrbitFlatTable to DataFormats/NanoAOD, as suggested by Matti
