+++ DataFormats/L1Scouting/BuildFile.xml
@@ -1,6 +1,5 @@
 <use name="FWCore/Utilities"/>
 <use name="DataFormats/Common"/>
-<use name="DataFormats/NanoAOD"/>
 <export>
   <lib name="1"/>
 </export>
