+++ DataFormats/L1Scouting/src/classes.h
@@ -5,4 +5,3 @@
 #include "DataFormats/L1Scouting/interface/L1ScoutingMuon.h"
 #include "DataFormats/L1Scouting/interface/L1ScoutingCalo.h"
 #include "DataFormats/L1Scouting/interface/L1ScoutingBMTFStub.h"
-#include "DataFormats/L1Scouting/interface/OrbitFlatTable.h"
