+++ DataFormats/NanoAOD/src/classes.h
@@ -1,6 +1,7 @@
 #include "Rtypes.h"
 
 #include "DataFormats/NanoAOD/interface/FlatTable.h"
+#include "DataFormats/NanoAOD/interface/OrbitFlatTable.h"
 #include "DataFormats/NanoAOD/interface/MergeableCounterTable.h"
 #include "DataFormats/NanoAOD/interface/UniqueString.h"
 #include "DataFormats/Common/interface/Wrapper.h"
