+++ DataFormats/NanoAOD/src/classes_def.xml
@@ -44,5 +44,10 @@
         <version ClassVersion="3" checksum="2793726511"/>
     </class>
     <class name="edm::Wrapper<nanoaod::UniqueString>" />
+
+    <class name="l1ScoutingRun3::OrbitFlatTable" ClassVersion="3">
+        <version ClassVersion="3" checksum="1910072015"/>
+    </class>
+    <class name="edm::Wrapper<l1ScoutingRun3::OrbitFlatTable>"/>
 </lcgdict>
